backend/requirements.txt
fastapi
uvicorn
discord.py
psutil
httpx
openai
google-generativeai
python-dotenv
cryptography
